Rename some Python example scripts
It's not a great idea to use '-' in Python file names; this prevents
importing that file in any other code that may want to use its
functionality. This doesn't matter at all for these scripts since
they're all standalone and don't guard their execution with an
"if __name__ == '__main__'" check, but it's still a bad habit so
rename get-object.py to get_object.py.

It's also a bad idea to name a Python file the same thing as a standard
library module, which I found out the hard way when I mistyped "exit"
inside the python3 directory. Ubuntu's "command not found" helper script
(written in Python) imported this logging.py instead of the standard
library module, which resulted in a file named 'eixt' containing RETS
logs.